We did it! I think!
so this is after downloading the GitHub desktop client and connecting it to PyCharm community thingy

a bunch of files aside from main.py appeared and i don't know what they're for so that's cool

anyway, i'd like to just create a description or a readme or maybe just delete the whole thing. we'll see.